+++ spec/ergoterm/instance/focus_spec.lua
@@ -39,7 +39,7 @@ describe(".focus", function()
 
     focus(term)
 
-    assert.equal(term, collection.get_last_focused())
+    assert.equal(term, collection.get_target_for_bang())
   end)
 end)